+++ src/errors.spec.ts
@@ -1,4 +1,4 @@
-import { ServiceApiError, ServiceClientError } from './errors';
+import { CreateProfileApiError, CreateProviderApiError, ServiceApiError, ServiceClientError } from './errors';
 
 describe('errors', () => {
   describe('ServiceClientError', () => {
@@ -19,6 +19,32 @@ describe('errors', () => {
       });
 
       expect(error instanceof ServiceApiError).toBeTruthy();
+    });
+  });
+
+  describe('CreateProviderApiError', () => {
+    it('should work with instanceof operator', () => {
+      const error = new CreateProviderApiError({
+        status: 500,
+        title: 'Error',
+        detail: 'Test error',
+        instance: '/mocked',
+      });
+
+      expect(error instanceof CreateProviderApiError).toBeTruthy();
+    });
+  });
+
+  describe('CreateProfileApiError', () => {
+    it('should work with instanceof operator', () => {
+      const error = new CreateProfileApiError({
+        status: 500,
+        title: 'Error',
+        detail: 'Test error',
+        instance: '/mocked',
+      });
+
+      expect(error instanceof CreateProfileApiError).toBeTruthy();
     });
   });
 });