+++ src/errors.ts
@@ -1,4 +1,6 @@
 import { ServiceApiErrorResponse } from './interfaces';
+import { CreateProfileApiErrorResponse } from './interfaces/create_profile_api_error_response';
+import { CreateProviderApiErrorResponse } from './interfaces/create_provider_api_error_response';
 
 export class ServiceClientError extends Error {
   constructor(message: string) {
@@ -24,5 +26,33 @@ export class ServiceApiError extends ServiceClientError {
     this.instance = errorResponse.instance;
     this.title = errorResponse.title;
     this.detail = errorResponse.detail;
+  }
+}
+
+export class CreateProviderApiError extends ServiceApiError {
+  public providerJsonEquals?: boolean;
+  public validProviderNames?: string[];
+
+  constructor(errorResponse: CreateProviderApiErrorResponse) {
+    super(errorResponse);
+
+    Object.setPrototypeOf(this, CreateProviderApiError.prototype);
+
+    this.providerJsonEquals = errorResponse.provider_json_equals;
+    this.validProviderNames = errorResponse.valid_provider_names;
+  }
+}
+
+export class CreateProfileApiError extends ServiceApiError {
+  public contentIsEqual?: boolean;
+  public suggestedVersion?: string;
+
+  constructor(errorResponse: CreateProfileApiErrorResponse) {
+    super(errorResponse);
+
+    Object.setPrototypeOf(this, CreateProfileApiError.prototype);
+
+    this.contentIsEqual = errorResponse.content_is_equal;
+    this.suggestedVersion = errorResponse.suggested_version;
   }
 }