+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `createProvider` throws `CreateProviderApiError`
+- `createProfile` throws `CreateProfileApiError`
+
 
 ## [2.0.0] - 2022-04-12
 ### Changed